requirements.txt
kafka-python==2.0.2
pandas==2.2.2
scikit-learn==1.2.2
joblib==1.2.0
pytest==7.3.1
prometheus_client==0.18.0
flask==2.2.3
# apache-airflow==2.6.3
PyYAML==6.0
numpy<2.0 # Added to fix binary incompatibility with scikit-learn 1.2.2
